services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
prometheus-client==0.21.1
redis==5.2.1
pika==1.3.2
PyJWT==2.10.1
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1

